DEPLOY: Ensure any routes are visitable
Because this an SPA on the server side there should only be one
entry point i.e. index.html. But our routing happens client side so
all routes need to be pointed to the index.html. The _redirects file is
expected by Netlify and it enables us to point all routes to index.html