Fix the shuffle feature
The approach to implementing this feature was twofold:
1. force the wordGrid component to rerender by setting a key
on the component and changing that key when the 'Shuffle'
button is clicked.

2. Use a method instead of a computed property to get the
words for the words prop passed to the the WordGrid component.

This approach had the unintended side effect of rearranging the
words everytime the user submitted a guess. The words should only
rearrange when the user taps the 'Shuffle' button.

FIX: Revert to a computed property for the 'words' prop. Set a
new data property to the wordGridKey when the 'Shuffle' button is
clicked and generate a new key. By comparing the 2 keys, we decide if
a new order is required. The 2 keys will only be different if the
'Shuffle' button has been clicked and therefore the words need to
be "shuffled".